requirements.txt
langchain
langchain-openai
langchain_core
langchain-community
langchainhub
openai
ragas
tiktoken
cohere
faiss_cpu
python-dotenv
beautifulsoup4
unstructured